+++ userspace/tauxprint.c
@@ -21,7 +21,7 @@ main(void)
             break;
         }
 
-        if (ioctl(fd, TAUX_SET_LED_STR, (int)buf) < 0) {
+        if (ioctl(fd, TAUX_SET_LED_STR, (intptr_t)buf) < 0) {
             fprintf(stderr, "Cannot display that string!\n");
         }
     }
